# backend/OmnikeeperOnlineInboundAdapter/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Landscape Registry REST API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:44378"
    }
  ],
  "paths": {
...
    "/api/v{version}/CI/getAllCITypes": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "list of all CI-types",
        "operationId": "GetAllCITypes",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CITypeDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CITypeDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CITypeDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/v{version}/CI/getAllCIIDs": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "list of all CI-IDs",
        "operationId": "GetAllCIIDs",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/v{version}/CI/getCIByID": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "single CI by CI-ID",
        "operationId": "GetCIByID",
        "parameters": [
          {
            "name": "layerIDs",
            "in": "query",
            "description": "Specifies which layers contribute to the result, and in which order",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": "Specifies which layers contribute to the result, and in which order"
            }
          },
          {
            "name": "CIID",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "description": "",
              "format": "uuid"
            }
          },
          {
            "name": "atTime",
            "in": "query",
            "description": "Specify datetime, for which point in time to get the data; leave empty to use current time (https://www.newtonsoft.com/json/help/html/DatesInJSON.htm)",
            "schema": {
              "type": "string",
              "description": "Specify datetime, for which point in time to get the data; leave empty to use current time (https://www.newtonsoft.com/json/help/html/DatesInJSON.htm)",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CIDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CIDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CIDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/v{version}/CI/getCIsByID": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "multiple CIs by CI-ID\r\n!Watch out for the query URL getting too long because of a lot of CIIDs!\r\nTODO: consider using POST",
        "operationId": "GetCIsByID",
        "parameters": [
          {
            "name": "layerIDs",
            "in": "query",
            "description": "Specifies which layers contribute to the result, and in which order",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": "Specifies which layers contribute to the result, and in which order"
            }
          },
          {
            "name": "CIIDs",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              },
              "description": ""
            }
          },
          {
            "name": "atTime",
            "in": "query",
            "description": "Specify datetime, for which point in time to get the data; leave empty to use current time (https://www.newtonsoft.com/json/help/html/DatesInJSON.htm)",
            "schema": {
              "type": "string",
              "description": "Specify datetime, for which point in time to get the data; leave empty to use current time (https://www.newtonsoft.com/json/help/html/DatesInJSON.htm)",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CIDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/v{version}/CI/getCIIDsOfNonEmptyCIs": {
      "get": {
        "tags": [
          "CI"
        ],
        "summary": "",
        "operationId": "GetCIIDsOfNonEmptyCIs",
        "parameters": [
          {
            "name": "layerIDs",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": ""
            }
          },
          {
            "name": "atTime",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string",
              "description": "",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributeValueType": {
        "enum": [
          "Text",
          "MultilineText",
          "Integer",
          "JSON",
          "YAML"
        ],
        "type": "string"
      },
      "AttributeValueDTO": {
        "required": [
          "isArray",
          "type",
          "values"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/AttributeValueType"
          },
          "isArray": {
            "type": "boolean"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AttributeState": {
        "enum": [
          "New",
          "Changed",
          "Removed",
          "Renewed"
        ],
        "type": "string"
      },
      "CIAttributeDTO": {
        "required": [
          "ciid",
          "id",
          "name",
          "state",
          "value"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "readOnly": true
          },
          "value": {
            "$ref": "#/components/schemas/AttributeValueDTO"
          },
          "ciid": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "state": {
            "$ref": "#/components/schemas/AttributeState"
          }
        }
      },
...
      "CITypeDTO": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "CIDTO": {
        "required": [
          "attributes",
          "id",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "type": {
            "$ref": "#/components/schemas/CITypeDTO"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CIAttributeDTO"
            },
            "readOnly": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "authorizationUrl": "https://host.docker.internal:8443/auth/realms/landscape/protocol/openid-connect/auth",
            "tokenUrl": "https://host.docker.internal:8443/auth/realms/landscape/protocol/openid-connect/token",
            "scopes": { }
          }
        }
      }
    }
  }
}